Add suport for using Tilda as a desktop (aka transparent wallpaper) terminal.
Set the window hint allowing us to persist across go-to-desktop functionality
invokations on all ICCCM/EWMH compliant window managers. Configurable setting,
defaults to no.

This patch adds support for it the wizard as well and reorders the appearance
of the checkboxes. The window and workspace related settings are now separated
by a line and activating the "show on desktop" option blocks the
"show on all workspaces" since it is redundant in this case.

Signed-off-by: Sebastian Geiger <sbastig@gmx.net>