key_grabber.c: Temporarily set X window type hint to DOCK
This patch fixes the window movement for most window managers. Without this
most windowm managers (such as metacity, compiz or mutter and possibly others)
will prevent the window to programmatically move itself outside the visible
screen area. As a result it was only possible to see the pull animation
on a few window managers that did not have this constraint (e.g 'xfwm4').

This patch temporarily removes the control of the window manager from the
window so we can properly handle the move animation.